requirements.txt
fastapi
uvicorn
requests
pymongo
spacy
tqdm
pydantic
sentence-transformers
torch
Pillow
python-dotenv
PyPDF2
python-docx
pinecone
huggingface_hub


